Move and improve `ArrayData` comparison function
