+++ crates/utils/re_arrow_util/src/lib.rs
@@ -2,6 +2,8 @@
 
 mod arrays;
 mod batches;
+mod compare;
 
 pub use self::arrays::*;
 pub use self::batches::*;
+pub use self::compare::*;
